requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
cryptography==38.0.3
importlib-metadata==5.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.11.0
keyrings.alt==4.2.0
more-itertools==9.0.0
psycopg2-binary==2.9.5
pycparser==2.21
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
SecretStorage==3.3.3
six==1.16.0
tornado==6.1
tzdata==2022.6
tzlocal==4.2
zipp==3.10.0
